Moved coord_to_r from rotate_galaxy to fire_tools
